+++ fuse_constraints/src/normal_prior_pose_2d.cpp
@@ -51,7 +51,7 @@ bool NormalPriorPose2D::Evaluate(double const* const* parameters, double* residu
   fuse_core::Vector3d full_residuals_vector;
   full_residuals_vector[0] = parameters[0][0] - b_[0];                          // position x
   full_residuals_vector[1] = parameters[0][1] - b_[1];                          // position y
-  full_residuals_vector[2] = fuse_core::wrapAngle2D(parameters[1][0] - b_[2]);  // orientation
+  full_residuals_vector[2] = fuse_core::wrapAngle(parameters[1][0] - b_[2]);  // orientation
 
   // Scale the residuals by the square root information matrix to account for the measurement uncertainty.
   Eigen::Map<fuse_core::VectorXd> residuals_vector(residuals, num_residuals());