+++ fuse_core/include/fuse_core/util.h
@@ -111,7 +111,7 @@ static inline T getYaw(const T w, const T x, const T y, const T z)
  * @param[in/out] angle Input angle to be wrapped to the [-Pi, +Pi) range. Angle is updated by this function.
  */
 template <typename T>
-void wrapAngle2D(T& angle)
+void wrapAngle(T& angle)
 {
   // Define some necessary variations of PI with the correct type (double or Jet)
   static const T PI = T(M_PI);
@@ -128,10 +128,10 @@ void wrapAngle2D(T& angle)
  * @return The equivalent wrapped angle
  */
 template <typename T>
-T wrapAngle2D(const T& angle)
+T wrapAngle(const T& angle)
 {
   T wrapped = angle;
-  wrapAngle2D(wrapped);
+  wrapAngle(wrapped);
   return wrapped;
 }