+++ fuse_core/CHANGELOG.rst
@@ -115,7 +115,7 @@ Changelog for package fuse_core
   not returned, but set in an in/out argument.
   * Move getPositiveParam and other param related functions to
   parameter.h from util.h, and updated the ros/unit tests accordingly.
-  * Fix wrapAngle2D expected range to [-Pi, +Pi)
+  * Fix wrapAngle expected range to [-Pi, +Pi)
   Instead of (-Pi, +Pi], and update unit test to reflect that.
 * Fix thread issue with UUID generation (`#167 <https://github.com/locusrobotics/fuse/issues/167>`_)
   * Add a mutex lock to the random UUID generation. The STL random number generator is not thread-safe.